requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
et-xmlfile==1.0.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
